% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixglm.R
\name{checkInit}
\alias{checkInit}
\title{Check initial values for mixglm and generate new ones if necessary}
\usage{
checkInit(
  stateValModels,
  statePrecModels,
  inputData,
  setInit,
  errorModel,
  linkFunction,
  Nstates,
  genNew = FALSE,
  genFn = function(x) rnorm(1, 0, 2)
)
}
\arguments{
\item{stateValModels}{A formula describing the regression relationship between the mean
of the response variable and the covariates for all mixture components, or a list of formulas
with each element giving the regression relationship between the mean
of the response variable and the covariates for each mixture component (ordered according to intercept of the
mixture component on the y-axis).  The response variable must be given on the
left-hand side of the formula.}

\item{statePrecModels}{A formula describing the regression relationship between the variances
in the mixture components and the covariates, or a list of
formulas with each mixture component giving the regression relationship between the variance in the
mixture component and the covariates for each mixture component (ordered according to
intercept of the mixture component on the y-axis).}

\item{inputData}{A data frame containing the covariate information.}

\item{setInit}{List of initial values. It is assumed that for each parameter, there is a vector
of initial values with length corresponding to \code{Nstates}}

\item{errorModel}{A string specifying the error distribution to be used
in the mixglm. This can be: \code{"gaussian"},
\code{"gamma"}, \code{"beta"} or \code{"negbinomial"}.}

\item{linkFunction}{A string specifying the link function. This can be: \code{"identity"},
\code{"log"}, \code{"logit"}, \code{"probit"}, or \code{"cloglog"}.}

\item{Nstates}{An integer specifying number of components in the mixture.}

\item{genNew}{A logical value specifying if new initial values should be generated even if
\code{setInit} contains legitimate values.}

\item{genFn}{A function used to generate new initial values. It has one argument which is
the component rank and it should return one value. Note that for intercept of precision of beta
distribution, \code{genFn(1) + 6} is used.}
}
\value{
A list of initial values
}
\description{
This function checks initial values for mixglm and if they are illegitimate, it tries
to generate new ones.
}
\author{
Adam Klimes
}
\keyword{internal}
