% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixglm.R
\name{coef.mixglm}
\alias{coef.mixglm}
\title{Extract model coefficients from a mixture model}
\usage{
\method{coef}{mixglm}(object, digit = NULL, ...)
}
\arguments{
\item{object}{an object of class "mixglm"}

\item{digit}{integer specifying the number of decimal places to be used.
Use \code{"NULL"} for no rounding.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A List of matrices of posterior mean values for each parameter
}
\description{
Posterior mean values from mixglm
}
\author{
Adam Klimes
}
