% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleExtensions.R
\name{dbetabinStateValueMembership}
\alias{dbetabinStateValueMembership}
\title{Probability Density of a response variable from beta-binomial mixture (Beta-Binomial Error)}
\usage{
dbetabinStateValueMembership(
  x,
  stateVal,
  statePrec,
  stateProb,
  numTrials,
  log = 0
)
}
\arguments{
\item{x}{A scalar value containing the value of the response variable}

\item{stateVal}{A numeric vector containing the mean values of the mixture components}

\item{statePrec}{A numeric vector containing the precision (reciprocal of the variance) of the mixture components}

\item{stateProb}{A numeric vector containing the probabilities of belonging
to each mixture component (internally normalised to one)}

\item{numTrials}{A scalar containing the number of trials}

\item{log}{If \code{TRUE} then return the log density}
}
\value{
A scalar containing the probability density
}
\description{
Calculate the probability density of a response variable from beta-binomial mixture.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
