% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPmeans.R
\name{DPmeans}
\alias{DPmeans}
\title{Penalized kmeans algorithm
Reference:
Brian Kulis & Michael I. Jordan, Revisiting k-means: New Algorithms via Bayesian Nonparametrics, 2012}
\usage{
DPmeans(x, lambda, tol = 0.001)
}
\description{
Penalized kmeans algorithm
Reference:
Brian Kulis & Michael I. Jordan, Revisiting k-means: New Algorithms via Bayesian Nonparametrics, 2012
}
