% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixglm.R
\name{findMin}
\alias{findMin}
\title{Find positions of local minima in a vector}
\usage{
findMin(x, extremes = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{extremes}{logical indicating if first and last values should be considered}
}
\value{
A vector of positions of minima in x
}
\description{
Finds position of all local minima in a vector including start
  and end point. For flat minima (identical subsequent values), it denotes middle point
}
\author{
Adam Klimes
}
\keyword{internal}
