% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obj_val.R
\name{get_obj_val}
\alias{get_obj_val}
\title{Calculate objective function (gaussian only)}
\usage{
get_obj_val(
  beta_int,
  alpha_int,
  gamma_int,
  y_all,
  X_all,
  W_all,
  K,
  p_x,
  lambda1,
  lambda2,
  lambda3,
  l1_weights
)
}
\description{
Calculate objective function (gaussian only)
}
