% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixglm.R
\name{getMinMax}
\alias{getMinMax}
\title{Find local minima and maxima in mixglm}
\usage{
getMinMax(slices, threshold = 0)
}
\arguments{
\item{slices}{slices of stability landscape from sliceMixglm() function}

\item{threshold}{numerical value specifying threshold for marking minima
and maxima. Marked maxima/minima differ in scaled probability density by
threshold value.}
}
\value{
A list containing the following components:
\item{\code{tipStable}}{ A list of lists of dataframes with
  tipping points (state == 0) and stable states (state == 1), categorized
  based on satifying the threshold (catSt == 1), with their scaled [0-1]
  probability density and response variable value}
\item{\code{mats}}{ An array of stability landscapes}
\item{\code{matsSt}}{ A list of scaled stability landscapes}
}
\description{
Finds local minima and maxima in slices of stability landscape
  from mixglm which represent multimodality over
  specified threshold.
}
\author{
Adam Klimes
}
\keyword{internal}
