% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_glm.R
\name{mix_glm}
\alias{mix_glm}
\title{Fit generalized linear models for integrated data}
\usage{
mix_glm(
  y_var,
  x_vars,
  w_vars = NULL,
  dat_list,
  family = c("gaussian", "binomial"),
  kappa = 1,
  lambda1 = 0,
  lambda2 = 0,
  lambda3 = 0,
  weighted = 0,
  penalty.factor = 1,
  coeff_group_numbers = NULL,
  beta_int = NULL,
  alpha_int = NULL,
  tol1 = 1e-05,
  tol2 = 1e-05,
  iter_max = 500,
  check = 0
)
}
\arguments{
\item{y_var}{response variable}

\item{x_vars}{predictors with heterogeneous covariate effects}

\item{w_vars}{predictors with homogeneous covariate effects}

\item{dat_list}{a list of data sets by studies}

\item{family}{a description of the error distribution and link function to be used in the model}

\item{kappa}{parameter for argumented Lagrangian function}

\item{lambda1}{penalty parameter for L1 norm}

\item{lambda2}{penalty parameter for multi-directional L2 norm}

\item{lambda3}{parameter for the penalized K-means algorithm, leading to lambda3/lambda2 penalty paramter on the subgroup sizes}

\item{coeff_group_numbers}{a vector specifying numbers of subgroups within each predictor. If NULL, penalized K-means is applied.}

\item{beta_int}{initial values of heterogeneous effects}

\item{alpha_int}{initial values of homogeneous effects}

\item{sigma_int}{initial values of scale parameters}

\item{sigma_fixed}{logic}

\item{penalty_method}{penalty method}
}
\description{
Fit generalized linear models for integrated data
}
