% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_nu_gamma.R
\name{pen_obj_by_p}
\alias{pen_obj_by_p}
\title{Objective function of nu and gamma}
\usage{
pen_obj_by_p(
  nu_int_p,
  gamma_int_p,
  beta_int_p,
  Lamb_int_p,
  kappa,
  lambda2,
  lambda3
)
}
\description{
Objective function of nu and gamma
}
