% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixglm.R
\name{plot.mixglm}
\alias{plot.mixglm}
\title{Plot results of a mixture model}
\usage{
\method{plot}{mixglm}(
  x,
  form = NULL,
  byChains = TRUE,
  transCol = TRUE,
  addWAIC = FALSE,
  setCol = c("#1b9e77", "#d95f02", "#7570b3", "#e7298a", "#66a61e", "#e6ab02", "#a6761d",
    "#666666"),
  drawAxes = TRUE,
  SDmult = 1,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class "mixglm"}

\item{form}{formula, such as y ~ pred, specifying variables to be plotted. By
default the first predictor is plotted on x axis.}

\item{byChains}{logical value indicating whether to plot mixture components for each chain}

\item{transCol}{logical value indicating usage of transparent colours}

\item{addWAIC}{logical value indication display of WAIC in upper right corner of the plot}

\item{setCol}{vector of colours to be used for mixture components}

\item{drawAxes}{logical value indicating whether values should be marked on axes}

\item{SDmult}{scalar multiplying visualized standard deviation (to make
lines for small standard deviation visible)}

\item{xlab}{a label for the x axis, defaults to predictor name.}

\item{ylab}{a label for the y axis, defaults to response name.}

\item{...}{additional arguments passed to \link[graphics]{plot}}
}
\description{
This function plots results of a mixture model on the
current graphical device.
}
\author{
Adam Klimes
}
