% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixglm.R
\name{print.mixglm}
\alias{print.mixglm}
\title{Print a mixture model}
\usage{
\method{print}{mixglm}(x, ...)
}
\arguments{
\item{x}{an object of class "mixglm"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{x}
}
\description{
This function prints summary information about mixglm
}
\author{
Adam Klimes
}
