% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleExtensions.R
\name{rgammaStateValueMembership}
\alias{rgammaStateValueMembership}
\title{Simulate a Variable from a gamma mixture (Gamma Error)}
\usage{
rgammaStateValueMembership(n, stateVal, statePrec, stateProb)
}
\arguments{
\item{n}{An integer scalar containing the number of variables to simulate (NIMBLE currently
only allows n = 1)}

\item{stateVal}{A numeric vector containing the mean values of the mixture components}

\item{statePrec}{A numeric vector containing the precision (reciprocal of the
variance) of the mixture components}

\item{stateProb}{A numeric vector containing the probabilities of belonging
to each mixture component (internally normalised to one)}
}
\value{
A scalar containing the simulated response variable
}
\description{
Simulate a variable from a gamma mixture.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
