% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixglm.R
\name{rMixglm}
\alias{rMixglm}
\title{Randomly generate data from a mixture model}
\usage{
rMixglm(mod, n = 1, newdata = NULL)
}
\arguments{
\item{mod}{an object of class "mixglm"}

\item{n}{number of generated values per predictor value(s)}

\item{newdata}{dataframe of predictor values to be predicted for.
If not provided, prediction is done for modelled data.}
}
\value{
A matrix with a column for each \code{n}
}
\description{
Random generation of response variable from identified mixture model
}
\author{
Adam Klimes
}
