% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixglm.R
\name{setBUGSVariableName}
\alias{setBUGSVariableName}
\title{Change Variable Names to BUGS-Friendly Versions}
\usage{
setBUGSVariableName(inName, warn = FALSE)
}
\arguments{
\item{inName}{A character vector of variable names}

\item{warn}{A logical value specifying if warnings should be printed for any
change in a variable name}
}
\value{
A character vector of BUGS-compliant variable names
}
\description{
Not all potential variable names used in R can be used in BUGS code without
producing a syntax error.  This function changes a vector of input names and ensures that
they are valid BUGS variable names.
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
