% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixglm.R
\name{sliceMixglm}
\alias{sliceMixglm}
\title{Plot slice from mixglm}
\usage{
sliceMixglm(
  mod,
  form = NULL,
  value = 0,
  byChains = TRUE,
  xlab = names(mod$data),
  doPlot = TRUE,
  setCol = c("#1b9e77", "#d95f02", "#7570b3", "#e7298a", "#66a61e"),
  plotEst = TRUE,
  xaxis = TRUE,
  addEcos = FALSE,
  ecosTol = 0.1,
  samples = 1000,
  randomSample = NULL,
  getParsD = FALSE
)
}
\arguments{
\item{mod}{an object of class "mixglm"}

\item{form}{formula with one predictor specifying which variables to plot}

\item{value}{numeric vector of values of the preditor specified by
\code{"form"} where the slice is done or data.frame with values of predictors
in named columns}

\item{byChains}{logical value indicating if slice should be done for each chain separately}

\item{xlab}{string used as label for x-axis}

\item{doPlot}{logical value indicating if plotting should be done}

\item{setCol}{vector of colours to be used for visualization of estimated states}

\item{plotEst}{logical value indicating if estimated mixture components should be visualized}

\item{xaxis}{logical value indicating if values should be marked on x-axis}

\item{addEcos}{logical value indicating if observations within \code{"ecosTol"}
from \code{"value"} should be visualized on the stability curve}

\item{ecosTol}{scalar specifying range of predictor from the \code{"value"}
to select observations to be visualized}

\item{samples}{scalar specifying number of samples to be taken along predictor}

\item{randomSample}{integer specifying how many random samples from posterior
distribution to take instead of summary. Use \code{"NULL"} for summary.}

\item{getParsD}{logical value indicating if the output should be parameters
of distributions instead of the slice. Requires \code{doPlot} to be \code{FALSE}}
}
\value{
A list containing the following components:
\itemize{
\item{\code{chainX}}{One or more items of a matrix of probability density
values along the response with a column for each \code{value}. One item is
returned if \code{byChains} is not \code{TRUE}, otherwise one item for each
chain. If \code{getParsD} is \code{TRUE}, parameters for probability density
distributions are returned instead.}
\item{\code{densFun}}{If \code{getParsD} is \code{TRUE}, function for
calculation of probability density is returned.}
\item{\code{resp}}{A vector of response values for which stability curves are
evaluated}}
}
\description{
This function plots probability density for given predictor value
}
\author{
Adam Klimes
}
