% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixglm.R
\name{str.mixglm}
\alias{str.mixglm}
\title{Compactly Display the Structure of a mixglm object}
\usage{
\method{str}{mixglm}(object, max.level = 2, give.attr = FALSE, ...)
}
\arguments{
\item{object}{an object of class "mixglm"}

\item{max.level}{integer giving maximal level of nesting of displayed structures}

\item{give.attr}{logical indicating if attributes should be shown}

\item{...}{arguments passed to \link[base]{NextMethod}}
}
\description{
Compactly display the internal structure of a mixglm object
}
\author{
Adam Klimes
}
