% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixglm.R
\name{summary.mixglm}
\alias{summary.mixglm}
\title{Summarize a mixture model}
\usage{
\method{summary}{mixglm}(
  object,
  byChains = FALSE,
  digit = 4L,
  absInt = FALSE,
  randomSample = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class "mixglm"}

\item{byChains}{logical value indicating if the summary should be calculated
for each chain separately}

\item{digit}{integer specifying the number of decimal places to be used.
Use \code{"NULL"} for no rounding.}

\item{absInt}{logical value indicating if intercepts for state values should
be absolute (by default, they represent differences)}

\item{randomSample}{integer specifying how many random samples from posterior
distribution to take instead of summary. Use \code{"NULL"} for summary.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns a list with one item being data.frame of quantiles of
posterior distributions of parameters. If \code{byChains} is \code{TRUE},
a data.frame is returned for each chain.
}
\description{
This function calculates posterior quantiles of parameters of
a mixture model across all chains or for each chain separately
}
\examples{
\dontrun{
set.seed(10)
n <- 200
x <- rnorm(n)
group <- rbinom(n, 1, 0.5)
y <- rnorm(n, 1 + 0.5 * x * c(-1, 1)[group + 1], 0.1)
plot(y ~ x)
dat <- data.frame(x, y)

mod <- mixglm(
  stateValModels = y ~ x,
  stateProbModels = ~ x,
  statePrecModels = ~ x,
  inputData = dat,
  numStates = 2)
summary(mod)

# random samples of parameters from posterior distribution
summary(mod, randomSample = 5)}
}
\author{
Adam Klimes
}
